% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAGISurvey-package.R
\docType{package}
\name{AAGISurvey-package}
\alias{AAGISurvey}
\alias{AAGISurvey-package}
\title{AAGISurvey: Create AAGI Survey URLs and Analyse AAGI Survey Results}
\description{
Create 'AAGI' service and support survey URLs interactively or scripted for sharing with partners.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://AAGI-AUS.github.io/AAGISurvey}
  \item \url{https://github.com/aagi-aus/AAGISurvey}
  \item Report bugs at \url{https://github.com/AAGI-AUS/AAGISurvey/issues}
}

}
\author{
\strong{Maintainer}: Adam Sparks \email{adam.sparks@curtin.edu.au} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})

Authors:
\itemize{
  \item Adam Sparks \email{adam.sparks@curtin.edu.au} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})
  \item Rose Megirian \email{rose.megirian@curtin.edu.au}
}

Other contributors:
\itemize{
  \item Grains Research and Development Corporation (\href{https://ror.org/02xwr1996}{ROR}) (GRDC Project CUR2210-005OPX (AAGI-CU)) [funder, copyright holder]
}

}
\keyword{internal}
