% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_url.R
\name{build_url}
\alias{build_url}
\title{Create the URL string from user provided values}
\usage{
build_url(
  base,
  support_type,
  design_type,
  analysis_type,
  aagi_node,
  organisation_type
)
}
\arguments{
\item{base}{Character string of base URL.}

\item{support_type}{Character vector of support types.}

\item{design_type}{Character string of design type.}

\item{analysis_type}{Character string of analysis type.}

\item{aagi_node}{Character string of AAGI node.}

\item{organisation_type}{Character string of organisation type.}
}
\value{
The full URL string.
}
\description{
Create the URL string from user provided values for returning to the user.
}
\examples{
build_url(
  base = "https://curtin.au1.qualtrics.com/jfe/form/SV_eXLvfgMz58RktQa",
  support_type = "S_D",
  design_type = "D_SP",
  analysis_type = "A_BIO",
  aagi_node = "CU",
  organisation_type = "O_ACA"
)

}
