% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_to_clip.R
\name{copy_to_clip}
\alias{copy_to_clip}
\title{Copy an \R Object's Text to the OS's Clipboard}
\usage{
copy_to_clip(x)
}
\arguments{
\item{x}{An object to be written to the system clipboard.}
}
\value{
A URL copied (invisibly) to the OS's clipboard.
}
\description{
Copy an \R Object's Text to the OS's Clipboard
}
\examples{
copy_to_clip(x = "https://github.com/AAGI-AUS")
}
\keyword{internal}
