% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_survey_url.R
\name{create_survey_url}
\alias{create_survey_url}
\title{Create an AAGI Survey URL}
\usage{
create_survey_url(
  support_type = NULL,
  design_type = NULL,
  analysis_type = NULL,
  aagi_node = NULL,
  organisation_type = NULL
)
}
\arguments{
\item{support_type}{Character string specifying the type of support being
evaluated. You must choose one.
Options are:
\describe{
\item{"S_D"}{Experimental design support}
\item{"S_A"}{Analysis support}
}}

\item{design_type}{Character string specifying the type of experimental
design (required if "S_D" is selected for \code{support_type}). You must choose
one.'
Options are:
\describe{
\item{"D_SP"}{Small plot trial design}
\item{"D_OFE"}{On farm experiment design}
\item{"D_GH"}{Glasshouse experiment design}
\item{"D_GC"}{Growth chamber experiment design}
\item{"D_LAB"}{Lab experiment design}
\item{"D_OTHER"}{Other design type chosen by the recipient}
}}

\item{analysis_type}{Character string of analysis type. Required if
"S_A" is selected for \code{support_type}. Options are:
\describe{
\item{"A_SP"}{Small plot trial data analysis}
\item{"A_OFE"}{On farm experiment data analysis}
\item{"A_PRO"}{Protected environment experiment data analysis
(glasshouse or growth chamber)}
\item{"A_BIO"}{Bioinformatics or genetic data}
\item{"A_BRE"}{Breeding or selection program data}
\item{"A_ENV"}{Environmental or geospatial data}
\item{"A_IMG"}{Imagery data}
\item{"A_REC"}{Farm records or monitoring data}
\item{"A_OTHER"}{Other type to be selected by recipient}
}}

\item{aagi_node}{Character string of AAGI node. Options are:
\describe{
\item{"ANU"}{Australian National University}
\item{"AU"}{Adelaide University}
\item{"CU"}{Curtin University}
\item{"QDPI"}{Queensland Department of Primary Industries}
\item{"UQ"}{University of Queensland}
\item{"UWA"}{University of Western Australia}
}}

\item{organisation_type}{Character string of organisation type. Options are:
\describe{
\item{"O_GRO"}{Grower group, industry association or farm cooperative}
\item{"O_AGR"}{Agronomy or farm advisory practice}
\item{"O_ACA"}{Academic institution}
\item{"O_GOV"}{Government agency or department}
\item{"O_BRE"}{Seed or breeding company}
\item{"O_TEC"}{Technology, biotechnology or chemical company}
\item{"O_OTHER"}{Other organisation type chosen by the recipient}
}}
}
\value{
Once the relevant information has been provided, the
function returns the completed survey URL, prints a short summary of the
details you supplied and copies the URL to your operating system clipboard
for inclusion in an email or other communication.
}
\description{
This function is designed to be used each time you need to
generate a survey URL for requesting feedback from a partner. It embeds
information already known by the team so that respondents do not need to
provide it. This includes the type of support provided, the output delivered,
the AAGI node that delivered the work, and the organisation type of the
partner. These details are included as metadata within the URL so that the
survey platform can automatically link each respondents feedback to the
relevant contextual information.
}
\details{
You can use the function in two ways. You can supply all required arguments
directly in your script, or you can run the function in interactive mode. In
interactive use, the function presents a menu in the console that guides you
through each field.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# create a survey URL for
# - a small plot trial design
# - for a government agency or department
# - performed by Curtin node
create_survey_url(
  support_type = "S_D",
  design_type = "D_SP",
  aagi_node = "CU",
  organisation_type = "O_GOV"
)

# create a survey URL for
# - a bioinformatics analysis
# - for an academic institution
# - performed by Adelaide University node
create_survey_url(
  support_type = "S_A",
  analysis_type = "A_BIO",
  aagi_node = "AU",
  organisation_type = "O_ACA"
)
\dontshow{\}) # examplesIf}
}
\author{
Rose Megirian, \email{rose.megirian@curtin.edu.au} and Adam H.
Sparks, \email{adam.sparks@curtin.edu.au}
}
