% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_survey_url.R
\name{ensure_valid}
\alias{ensure_valid}
\title{Validate user inputs}
\usage{
ensure_valid(vals, allowed, field)
}
\arguments{
\item{vals}{Character vector of user inputs.}

\item{allowed}{Named character vector of allowed values.}

\item{field}{Character string of field name for error messages.}
}
\value{
\code{NULL} invisibly if all values are valid, otherwise throws an
informative error.
}
\description{
Validate user inputs
}
\keyword{internal}
