% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_survey_url.R
\name{pick_codes}
\alias{pick_codes}
\title{Pick and match codes from a named character vector}
\usage{
pick_codes(dict, title, multiple = FALSE)
}
\value{
A character vector of the selected codes.
}
\description{
Ensures that all necessary values are provided and prompts for a response if
they are not.
}
\keyword{internal}
