% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAGISurvey-package.R
\docType{package}
\name{AAGISurvey-package}
\alias{AAGISurvey}
\alias{AAGISurvey-package}
\title{AAGISurvey: Create AAGI Survey URLs and Analyse AAGI Survey Results}
\description{
Create 'AAGI' service and support survey URLs interactively or scripted for sharing with partners.
}
\author{
\strong{Maintainer}: Adam Sparks \email{adam.sparks@curtin.edu.au}

Authors:
\itemize{
  \item Rose Megirian \email{rose.megirian@curtin.edu.au}
}

}
\keyword{internal}
